Documentation of SubmodularUtilities.jl

Documentation of SubmodularUtilities.jl

Maximization

This subset contains functions related to submodular maximization.

Lazy Greedy

SubmodularUtilities.lazy_greedy — Method.

lazy_greedy(f, ground_set, k)

ground_set should be an array of unique integers, which denotes the ground set of monotone submodular function f. f represents a monotone submodular function. If T is an array whose elements constitute a subset of ground_set, f(T) is the function value of T. k is the cardinality constraint. lazy_greedy outputs an array of length k using the lazy greedy algorithm.

source

Constraint

SubmodularUtilities.Constraint — Type.

Constraint

The constraints are A*x sense b and lb <= x <= ub. sense is a vector of constraint sense characters <, =, and >. A is a 2-dimensional array, which denotes the constraint matrix. b is the right-hand side vector.

source

Meta-Frank-Wolfe with Variance Reduction

The Meta-Frank-Wolfe algorithm solves the online maximization problem of monotone continuous DR-submodular functions. The vanilla Meta-Frank-Wolfe algorithm is proposed in [CHK] and works when accurate gradient is available. The verison with variance reduction is proposed in [CHHK] and solves the problem when only an unbiased estimate of gradient is available.

[CHK]
Lin Chen, Hamed Hassani, and Amin Karbasi, “Online Continuous Submodular Maximization”, in Proc. of AISTATS 2018.

[CHHK]
Lin Chen, Christopher Harshaw, Hamed Hassani, and Amin Karbasi, “Projection-Free Online Optimization with Stochastic Gradient: From Convexity to Submodularity“, in Proc. of ICML 2018.

SubmodularUtilities.meta_frank_wolfe — Method.

meta_frank_wolfe(dim_x, f_list, g_list, eta, K, constraint; variance_reduction = false, n0 = rand(dim_x))

Args:

dim_x: The dimension of the variables that we optimize.

f_list: List of continuous DR-submodular functions [f_1, f_2, ..., f_T].

g_list: List of (unbiased estimates of) the gradients of functions in f_list.

eta: This implementation uses Follow-Perturbed-Leader (FPL) for linear losses (which can be found in Hazan's book Introduction to Online Convex Optimization) as the online linear maximization algorithm. eta is the parameter of FPL.

K: Number of instances of the chosen online linear maximization algorithm.

constraint: Constraint set. It is an object of type Constraint.

variance_reduction: Boolean variable which is true if variance reduction is enabled

n0: the value of n0 in FPL for linear losses.

Returns:

Array of rewards [f_1(x_1), f_2(x_2), ..., f_T(f_T)].

source

Projection Operator

SubmodularUtilities.get_projection_operator — Method.

get_projection_operator(constraint; dim_x = size(constraint.A)[2])

Args:

constraint: Constraint set. It is an object of type Constraint.

dim_x: the dimension of the point. The default value is deduced from the size of constraint.A.

Returns: A function projection such that projection(x) outputs the projection of x onto the set constraint.

source

Online Gradient Ascent

SubmodularUtilities.online_gradient_ascent — Method.

online_gradient_ascent(x0, f_list, g_list, projection; step_size = t->1/t)

Args:

x0: the initial value

f_list: List of continuous DR-submodular functions [f_1, f_2, ..., f_T].

g_list: List of (unbiased estimates of) the gradients of functions in f_list.

projection: Projection operator of the constraint set. You may want to get it by calling get_projection_operator.

step_size: A function such that step_size(t) is the step size at the t-th iteration.

Returns:

Array of rewards [f_1(x_1), f_2(x_2), ..., f_T(f_T)].

source

Rounding

This subset contains functions related to rounding a fractional vector between 0 and 1 into a binary vector (equivalently, a discrete set).

Pipage Rounding

SubmodularUtilities.pipage_round — Method.

pipage_round(x)

Given an array x whose every entry is between 0 and 1, round x into a binary vector using the pipage rounding algorithm presented in [ccpv].

[ccpv]
Calinescu, Gruia, et al. "Maximizing a monotone submodular function subject to a matroid constraint." SIAM Journal on Computing 40.6 (2011): 1740-1766.

source

Random Rounding

SubmodularUtilities.random_round — Method.

random_round(x)

Given an array x whose every entry is between 0 and 1, round x into a binary vector. The i-th entry of the output vector is 1 with probability x[i] and is 0 otherwise.

source

Multilinear Extension

This subset contains functions related to multilinear extension of submodular set functions.

Random Evaluation of Multilinear Extension

SubmodularUtilities.get_random_evaluation_of_multilinear_extension — Function.

get_random_evaluation_of_multilinear_extension(f_discrete)

The input f_discrete is assumed to be a monotone submodular function defined on the ground set 1:n, where n is some positive integer. Given an array T whose elements are chosen from 1:n, f_discrete(T) is the function value of the subset T. get_random_evaluation_of_multilinear_extension returns a function that takes a size-n array x as input and outputs an unbiased estimate of the function value of the multilinear extension of f_discrete at x. In other words, the output is a function that maps x to f_discrete(random_round(x)).

source

Random Gradient of Multilinear Extension

SubmodularUtilities.get_random_gradient_of_multilinear_extension — Function.

get_random_gradient_of_multilinear_extension(f_discrete)

The input f_discrete is assumed to be a monotone submodular function defined on the ground set 1:n, where n is some positive integer. Given an array T whose elements are chosen from 1:n, f_discrete(T) is the function value of the subset T. get_random_gradient_of_multilinear_extension returns a function that takes a size-n array x as input and outputs an unbiased estimate of the gradient of the multilinear extension of f_discrete at x.

In other words, the output function first computes S = random_round(x). Then the i-th partial derivative is f_discrete(vcat(S, i)) - f_discrete(setdiff(S, i)).

source

Submodular Functions

This subset contains several examples of submodular functions. Given input data, the functions named get_function_... below return a submodular function.

Exemplar-Based Clustering

SubmodularUtilities.get_function_exemplar_based_clustering — Function.

get_function_exemplar_based_clustering(data)

Returns the objective function of exemplar-based clustering defined on the ground set 1:size(data)[1]. The argument data is a 2-dimensional array. Each row represents a data point and each column represents an attribute.

source

Active Set Selection

SubmodularUtilities.get_function_active_set_selection — Function.

get_function_active_set_selection(data; sigma = 1., h = 0.75)

Returns the objective function of active set selection defined on the ground set 1:size(data)[2]. The argument data is a 2-dimensional array. Each row represents a data point and each column represents an attribute. The covariance matrix uses a Gaussian kernel exp(-norm(data[:, i] - data[:, j])^2 / h^2). The function value of the objective function evaluated at the subset S is 0.5 * log(det(eye(length(S)) + cov_matrix[S, S] / sigma^2)).

source